change delete host modal and clean up pluralize util
