+++ frontend/pages/hosts/components/DeleteHostModal/DeleteHostModal.tsx
@@ -1,5 +1,7 @@
 import React from "react";
 
+import strUtils from "utilities/strings";
+
 import Modal from "components/Modal";
 import Button from "components/buttons/Button";
 import CustomLink from "components/CustomLink";
@@ -29,11 +31,18 @@ const DeleteHostModal = ({
   hostName,
   isUpdating,
 }: IDeleteHostModalProps): JSX.Element => {
+  const pluralizeHost = () => {
+    if (!selectedHostIds) {
+      return "host";
+    }
+    return strUtils.pluralize(selectedHostIds.length, "host", "s", "");
+  };
+
   const hostText = () => {
     if (selectedHostIds) {
       return `${selectedHostIds.length}${
         isAllMatchingHostsSelected ? "+" : ""
-      } ${selectedHostIds.length === 1 ? "host" : "hosts"}`;
+      } ${pluralizeHost()}`;
     }
     return hostName;
   };
@@ -58,17 +67,18 @@ const DeleteHostModal = ({
     >
       <>
         <p>
-          This action will delete <b>{hostText()}</b> from your Fleet instance.
-          {largeVolumeText()}
+          This will remove the record of <b>{hostText()}</b>.{largeVolumeText()}
         </p>
-        <p>If the hosts come back online, they will automatically re-enroll.</p>
         <p>
-          To prevent re-enrollment,{" "}
+          The {pluralizeHost()} will re-appear unless fleet&apos;s agent is
+          uninstalled.
+        </p>
+        <p>
           <CustomLink
             url={
               "https://fleetdm.com/docs/using-fleet/faq#how-can-i-uninstall-the-osquery-agent"
             }
-            text={"uninstall the osquery agent"}
+            text={"Uninstall Fleet's agent"}
             newTab
           />
         </p>
